raylib: support wayland by making xorg dependency optional
When building raylib on a wayland linux platform, there is no reason to
depend on xorg libraries.

Also add GLES 3.0 option, support for this was added in this commit:
https://github.com/raysan5/raylib/commit/3b09ef58ca03426ef1572ae783c9a2f2ea373ab8

And make the test output more explicit, as it was not clear what "unit
circle colides with itself" was supposed to mean. It seems to be the
successful case.